--- src/core/environment/model/index.ts
@@ -1,1 +1,0 @@
-export * from './environment.model';
