--- src/core/environment/service/index.ts
@@ -1,1 +1,0 @@
-export * from './environment.service';
