+++ src/core/flag-manager/model/environment.model.ts
