+++ src/core/flag-manager/model/index.ts
@@ -1,2 +1,3 @@
 export * from './category.model';
 export * from './flag.model';
+export * from './environment.model';
